refactor: 숙소 서비스에 CQRS 적용 수정
