+++ src/main/java/com/backoffice/upjuyanolja/domain/accommodation/controller/AccommodationBackofficeController.java
@@ -1,34 +1,22 @@
 package com.backoffice.upjuyanolja.domain.accommodation.controller;
 
 import com.backoffice.upjuyanolja.domain.accommodation.dto.request.AccommodationRegisterRequest;
-import com.backoffice.upjuyanolja.domain.accommodation.dto.response.AccommodationDetailResponse;
 import com.backoffice.upjuyanolja.domain.accommodation.dto.response.AccommodationInfoResponse;
 import com.backoffice.upjuyanolja.domain.accommodation.dto.response.AccommodationOwnershipResponse;
-import com.backoffice.upjuyanolja.domain.accommodation.dto.response.AccommodationPageResponse;
-import com.backoffice.upjuyanolja.domain.accommodation.dto.response.ImageResponse;
 import com.backoffice.upjuyanolja.domain.accommodation.service.usecase.AccommodationCommandUseCase;
+import com.backoffice.upjuyanolja.domain.accommodation.service.usecase.AccommodationQueryUseCase;
 import com.backoffice.upjuyanolja.global.security.SecurityUtil;
 import jakarta.validation.Valid;
-import java.time.LocalDate;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.hibernate.validator.constraints.Length;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.web.PageableDefault;
-import org.springframework.format.annotation.DateTimeFormat;
-import org.springframework.format.annotation.DateTimeFormat.ISO;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.multipart.MultipartFile;
 
 @Slf4j
 @RestController
@@ -38,6 +26,7 @@ import org.springframework.web.multipart.MultipartFile;
 public class AccommodationBackofficeController {
 
     private final AccommodationCommandUseCase accommodationCommandUseCase;
+    private final AccommodationQueryUseCase accommodationQueryUseCase;
     private final SecurityUtil securityUtil;
 
     @PostMapping
@@ -55,7 +44,7 @@ public class AccommodationBackofficeController {
     public ResponseEntity<AccommodationOwnershipResponse> getAccommodationOwnership() {
         log.info("GET /backoffice-api/accommodations");
 
-        AccommodationOwnershipResponse response = accommodationCommandUseCase
+        AccommodationOwnershipResponse response = accommodationQueryUseCase
             .getAccommodationOwnership(securityUtil.getCurrentMemberId());
         return ResponseEntity.status(HttpStatus.OK).body(response);
     }
