+++ src/main/java/com/backoffice/upjuyanolja/domain/accommodation/service/usecase/AccommodationQueryUseCase.java
@@ -1,32 +1,31 @@
 package com.backoffice.upjuyanolja.domain.accommodation.service.usecase;
 
-import com.backoffice.upjuyanolja.domain.accommodation.entity.Accommodation;
-import com.backoffice.upjuyanolja.domain.accommodation.entity.AccommodationImage;
+import com.backoffice.upjuyanolja.domain.accommodation.dto.response.AccommodationDetailResponse;
+import com.backoffice.upjuyanolja.domain.accommodation.dto.response.AccommodationOwnershipResponse;
+import com.backoffice.upjuyanolja.domain.accommodation.dto.response.AccommodationPageResponse;
 import com.backoffice.upjuyanolja.domain.accommodation.entity.AccommodationOption;
-import com.backoffice.upjuyanolja.domain.accommodation.entity.AccommodationOwnership;
 import com.backoffice.upjuyanolja.domain.accommodation.entity.Address;
 import com.backoffice.upjuyanolja.domain.accommodation.entity.Category;
-import com.backoffice.upjuyanolja.domain.member.entity.Member;
-import java.util.List;
+import java.time.LocalDate;
 import lombok.Builder;
+import org.springframework.data.domain.Pageable;
 
 public interface AccommodationQueryUseCase {
 
-    Accommodation saveAccommodation(AccommodationSaveRequest accommodationSaveRequest);
+    AccommodationPageResponse findAccommodations(
+        String Category,
+        boolean onlyHasCoupon,
+        String keyword,
+        Pageable pageable
+    );
 
-    Accommodation getAccommodationById(long id);
+    AccommodationDetailResponse findAccommodationWithRooms(
+        Long accommodationId,
+        LocalDate startDate,
+        LocalDate endDate
+    );
 
-    boolean existsById(Long accommodationId);
-
-    AccommodationOwnership saveOwnership(Member member, Accommodation accommodation);
-
-    List<AccommodationOwnership> getOwnershipByMember(Member member);
-
-    boolean existsOwnershipByMemberAndAccommodation(Member member, Accommodation accommodation);
-
-    Category getCategoryByName(String name);
-
-    List<AccommodationImage> saveAllImages(List<AccommodationImage> images);
+    AccommodationOwnershipResponse getAccommodationOwnership(long memberId);
 
     @Builder
     record AccommodationSaveRequest(
