+++ src/main/java/com/backoffice/upjuyanolja/domain/point/service/PointService.java
@@ -2,6 +2,7 @@ package com.backoffice.upjuyanolja.domain.point.service;
 
 
 import com.backoffice.upjuyanolja.domain.accommodation.service.AccommodationCommandService;
+import com.backoffice.upjuyanolja.domain.accommodation.service.AccommodationQueryService;
 import com.backoffice.upjuyanolja.domain.coupon.entity.Coupon;
 import com.backoffice.upjuyanolja.domain.coupon.entity.CouponIssuance;
 import com.backoffice.upjuyanolja.domain.coupon.exception.InsufficientPointsException;
@@ -76,6 +77,7 @@ public class PointService {
     private final PointUsageRepository pointUsageRepository;
 
     private final AccommodationCommandService accommodationCommandService;
+    private final AccommodationQueryService accommodationQueryService;
     private final CouponIssuanceGetService couponIssuanceGetService;
     private final MemberGetService memberGetService;
 
@@ -198,7 +200,7 @@ public class PointService {
             pointTotalDetailResponses.size());
 
         for (int i = startIndex; i < endIndex; i++) {
-            result.add(PointTotalDetailResponse.from(i+1, pointTotalDetailResponses.get(i)));
+            result.add(PointTotalDetailResponse.from(i + 1, pointTotalDetailResponses.get(i)));
         }
 
         long total = pointTotalDetailResponses.size();
@@ -345,6 +347,7 @@ public class PointService {
         point.updatePointBalance(totalBalance);
         pointRepository.save(point);
     }
+
     private void updateChargePointStatus(PointCharges pointCharges, PointStatus pointStatus) {
         pointCharges.updatePointStatus(pointStatus);
         pointCharges.updateRefundable(false);
@@ -472,7 +475,7 @@ public class PointService {
                         createCouponIssuancesMap(couponIssuances);
                     CouponIssuance selectCouponIssuance = couponIssuances.get(0);
                     String accommodationName =
-                        accommodationCommandService
+                        accommodationQueryService
                             .findAccommodationByRoomId(selectCouponIssuance.getRoom().getId())
                             .getName();
 
@@ -661,7 +664,7 @@ public class PointService {
         long correctBalance =
             Optional.ofNullable(pointChargesRepository.sumTotalPaidPoint(point)).orElse(0L) +
                 Optional.ofNullable(pointChargesRepository.sumTotalRemainedPoint(point)).orElse(0L);
-        if(totalBalance != correctBalance){
+        if (totalBalance != correctBalance) {
             throw new PointTradeFailedException();
         }
     }
